Multi dateformat (#1)
v1.1.0

Adds multi-format support.

dateFormat and timeFormat now accept arrays.

Whenever parsing a date, every entry in each array is considered. However, when formatting a date, only the first entry is ever used.

The outcome of this change is that given dateFormat=['DD/MM/YYYY', 'DDMMYYYY'], a user is able to type both "31/12/2020" and "31122020" with the same outcome. In the second case, where the secondary format is used, once a valid date is entered it will automatically convert to the primary date.

This PR also makes the following misc. changes:
- change package.json homepage info to point to this new repo
- add monkey-patch to allow running playground in node 20 (does not affect actual build)
- allow react 18 as peer dependency (this is the only change since we forked from upstream)
- add a yarn lock (for dev purposes)